Chat_bot/requirements.txt
python-telegram-bot[job-queue]>=21.0
google-genai>=1.0.0
aiosqlite>=0.20.0
PyGithub>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
